rag_service/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
qdrant-client==1.17.1
redis==5.0.1
openai==1.3.9
httpx==0.25.2
structlog==23.2.0
python-docx==1.1.2
PyPDF2==3.0.1
openpyxl==3.1.5
pytesseract==0.3.10
Pillow==10.1.0
librosa==0.10.0
sentence-transformers==2.2.2
